final version update
